Reorganise the structure of UTTT pages.
