+++ aisoc/client/src/competitions/uttt/pages/Game/index.js
@@ -1,0 +1,3 @@
+import Game from './Game';
+
+export default Game;
