+++ aisoc/client/src/competitions/uttt/pages/Home/RulesTab.js
