+++ aisoc/client/src/competitions/uttt/pages/Home/SubmissionGuideTab.scss
