+++ aisoc/client/src/competitions/uttt/pages/Match/index.js
@@ -1,0 +1,3 @@
+import Match from './Match';
+
+export default Match;
