+++ aisoc/client/src/competitions/uttt/pages/User/User.js
@@ -3,8 +3,8 @@ import LoadingPlaceholder from 'components/LoadingPlaceholder';
 import PairMatches from 'components/PairMatches';
 import { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
-import UTTTAPI from '../api';
-import MatchRow from '../components/MatchRow';
+import UTTTAPI from '../../api';
+import MatchRow from '../../components/MatchRow';
 
 export default function User({ baseUrl }) {
   const { user } = useParams();
