+++ aisoc/client/src/competitions/uttt/pages/User/index.js
@@ -1,0 +1,3 @@
+import User from './User';
+
+export default User;
